DRY up walking the Settings by moving into a module
