+++ lib/vmdb/settings.rb
@@ -2,11 +2,11 @@ require 'config'
 require_dependency 'patches/config_patch'
 require_dependency 'vmdb/settings/database_source'
 require_dependency 'vmdb/settings/hash_differ'
+require_dependency 'vmdb/settings/walker'
 
 module Vmdb
   class Settings
-    # if you change, please also change over in tools/fix_auth/models.rb
-    PASSWORD_FIELDS = %i(bind_pwd password amazon_secret).to_set.freeze
+    PASSWORD_FIELDS = Vmdb::Settings::Walker::PASSWORD_FIELDS
     DUMP_LOG_FILE   = Rails.root.join("log/last_settings.txt").freeze
 
     cattr_accessor :last_loaded
@@ -27,23 +27,8 @@ module Vmdb
       activate
     end
 
-    # if you change, please also change over in tools/fix_auth/auth_config_model.rb
     def self.walk(settings = ::Settings, path = [], &block)
-      settings.each do |key, value|
-        new_path = path.dup << key
-
-        yield key, value, new_path, settings
-
-        case value
-        when settings.class
-          walk(value, new_path, &block)
-        when Array
-          value.each_with_index do |v, i|
-            walk(v, new_path.dup << i, &block) if v.kind_of?(settings.class)
-          end
-        end
-      end
-      settings
+      Walker.walk(settings, path, &block)
     end
 
     def self.activate
@@ -77,15 +62,15 @@ module Vmdb
     end
 
     def self.mask_passwords!(settings)
-      walk_passwords(settings) { |k, _v, h| h[k] = "********" }
+      Walker.mask_passwords!(settings)
     end
 
     def self.decrypt_passwords!(settings)
-      walk_passwords(settings) { |k, v, h| h[k] = MiqPassword.try_decrypt(v) }
+      Walker.decrypt_passwords!(settings)
     end
 
     def self.encrypt_passwords!(settings)
-      walk_passwords(settings) { |k, v, h| h[k] = MiqPassword.try_encrypt(v) }
+      Walker.encrypt_passwords!(settings)
     end
 
     def self.dump_to_log_directory(settings)
@@ -151,13 +136,6 @@ module Vmdb
       Kernel.const_set(::Config.const_name, settings)
     end
     private_class_method :reset_settings_constant
-
-    def self.walk_passwords(settings)
-      walk(settings) do |key, value, _path, owning|
-        yield(key, value, owning) if value.present? && PASSWORD_FIELDS.include?(key.to_sym)
-      end
-    end
-    private_class_method :walk_passwords
 
     def self.apply_settings_changes(resource, deltas)
       resource.transaction do
